Skills/requirements.txt
# AI Employee System Python Dependencies
# Install with: pip install -r requirements.txt

# Google API client libraries (for Gmail Watcher and Email MCP Server)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Configuration and utilities
pyyaml>=6.0.1
html2text>=2020.1.16
python-dotenv>=1.0.0

# Optional: for better logging
colorlog>=6.7.0

# LinkedIn MCP Server dependencies
fastmcp>=3.0.0
inquirer>=3.4.0
patchright>=1.40.0

# Playwright for browser automation
playwright>=1.40.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0
